ISPN-11463 Cache manager stats are mistakenly exported as MP metrics even when stats disabled
